+++ cloudinary-core/src/main/java/com/cloudinary/EagerTransformation.java
@@ -2,10 +2,12 @@ package com.cloudinary;
 
 import com.cloudinary.utils.StringUtils;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
 public class EagerTransformation extends Transformation<EagerTransformation> {
+    protected String format;
 
     @SuppressWarnings("rawtypes")
     public EagerTransformation(List<Map> transformations) {
@@ -16,10 +18,40 @@ public class EagerTransformation extends Transformation<EagerTransformation> {
         super();
     }
 
+    public EagerTransformation format(String format) {
+        this.format = format;
+        return this;
+    }
+
+    public String getFormat() {
+        return format;
+    }
+
     @Override
-    protected void addFormat(List<String> components) {
-        if (StringUtils.isNotBlank(format)){
+    public String generate(Iterable<Map> optionsList) {
+        List<String> components = new ArrayList<String>();
+        for (Map options : optionsList) {
+            if (options.size() > 0) {
+                components.add(super.generate(options));
+            }
+        }
+
+        if (format != null){
             components.add(format);
         }
+
+        return StringUtils.join(components, "/");
+    }
+
+    @Override
+    public String generate(Map options) {
+        List<String> eager = new ArrayList<String>();
+        eager.add(super.generate(options));
+
+        if (format != null){
+            eager.add(format);
+        }
+
+        return StringUtils.join(eager, "/");
     }
 }
