+++ cloudinary-core/src/main/java/com/cloudinary/Transformation.java
@@ -628,23 +628,6 @@ public class Transformation<T extends Transformation> implements Serializable {
         return generate();
     }
 
-    protected String format;
-
-    public Transformation format(String format) {
-        this.format = format;
-        return this;
-    }
-
-    public String getFormat() {
-        return format;
-    }
-
-    protected void addFormat(List<String> components) {
-        if (format != null){
-            components.add(format);
-        }
-    }
-
     public String generate(Iterable<Map> optionsList) {
         List<String> components = new ArrayList<String>();
         for (Map options : optionsList) {
@@ -652,7 +635,6 @@ public class Transformation<T extends Transformation> implements Serializable {
                 components.add(generate(options));
             }
         }
-        addFormat(components);
         return StringUtils.join(components, "/");
     }
